Changed PreValidating Bot into PostValidating Initializer.
Changed Page Initializer to raise AutomationFailed.
